Change type of Piece and MovePoints size type to short for smaller MoveInfo size and better speed

+++ src/libpentobi_base/Piece.h
@@ -49,7 +49,7 @@ private:
 
     static const unsigned short value_uninitialized = range;
 
-    unsigned int m_i;
+    unsigned short m_i;
 
     bool is_initialized() const;
 };
